refactor remaining gear
