+++ source/gear/_gearDictionary.js
@@ -69,46 +69,19 @@ for (const file of [
 	"spiked-shield.js",
 	"steam-wall.js",
 	"stick.js",
-	"swordofthesun-base.js",
-	"swordofthesun-thiefs.js",
-	"swordofthesun-lethal.js",
-	"tempestuouswrath-base.js",
-	"tempestuouswrath-flanking.js",
-	"tempestuouswrath-opportunists.js",
-	"tornadoformation-base.js",
-	"tornadoformation-charging.js",
-	"tornadoformation-supportive.js",
-	"trident-base.js",
-	"trident-kinetic.js",
-	"trident-staggering.js",
-	"universalsolution-base.js",
-	"universalsolution-centering.js",
-	"universalsolution-tormenting.js",
-	"vacuumimplosion-base.js",
-	"vacuumimplosion-shattering.js",
-	"vacuumimplosion-urgent.js",
-	"vengefulvoid-base.js",
-	"vengefulvoid-hexing.js",
-	"vengefulvoid-numbing.js",
-	"warcry-base.js",
-	"warcry-flanking.js",
-	"warcry-weakening.js",
-	"warhammer-base.js",
-	"warhammer-fatiguing.js",
-	"warhammer-toxic.js",
-	"watersstillness-accelerating.js",
-	"watersstillness-base.js",
-	"watersstillness-cleansing.js",
-	"wavecrash-base.js",
-	"wavecrash-disenchanting.js",
-	"wavecrash-fatiguing.js",
-	"windburst-base.js",
-	"windburst-inspiring.js",
-	"windburst-toxic.js",
-	"wolfring-accurate.js",
-	"wolfring-base.js",
-	"wolfring-powerful.js",
-	"wolfring-swift.js"
+	"sword-of-the-sun.js",
+	"tempestuous-wrath.js",
+	"tornado-formation.js",
+	"trident.js",
+	"universal-solution.js",
+	"vacuum-implosion.js",
+	"vengeful-void.js",
+	"war-cry.js",
+	"warhammer.js",
+	"waters-stillness.js",
+	"wave-crash.js",
+	"wind-burst.js",
+	"wolf-ring.js"
 ]) {
 	/** @type {GearFamily} */
 	const gearFamily = require(`./${file}`);
