--- source/gear/watersstillness-accelerating.js
@@ -1,32 +1,0 @@
-const { GearTemplate } = require('../classes');
-const { ESSENCE_MATCH_STAGGER_ALLY } = require('../constants');
-const { concatTeamMembersWithModifier, changeStagger, generateModifierResultLines, combineModifierReceipts, addModifier } = require('../util/combatantUtil');
-const { joinAsStatement } = require('../util/textUtil');
-const { scalingSwiftness } = require('./shared/modifiers');
-
-module.exports = new GearTemplate("Accelerating Water's Stillness",
-	[
-		["use", "Relieve Stagger and grant <@{mod1Stacks}> @{mod1} to an ally and all allies with @{mod0}"],
-		["critical", "Stagger relieved x @{critBonus}"]
-	],
-	"Spell",
-	"Water"
-).setCost(350)
-	.setEffect((targets, user, adventure) => {
-		const { essence, modifiers: [targetModifier, swiftness], stagger, scalings: { critBonus } } = module.exports;
-		const allTargets = concatTeamMembersWithModifier(targets, user.team === "delver" ? adventure.delvers : adventure.room.enemies, targetModifier.name);
-		let pendingStaggerRelief = stagger;
-		if (user.essence === essence) {
-			pendingStaggerRelief += ESSENCE_MATCH_STAGGER_ALLY;
-		}
-		if (user.crit) {
-			pendingStaggerRelief *= critBonus;
-		}
-		changeStagger(allTargets, user, pendingStaggerRelief);
-		return [joinAsStatement(false, allTargets.map(target => target.name), "shrugs off", "shrug off", "some Stagger.")].concat(generateModifierResultLines(combineModifierReceipts(addModifier(allTargets, { name: swiftness.name, stacks: swiftness.stacks.calculate(user) }))));
-	}, { type: "single", team: "ally" })
-	.setSidegrades("Cleansing Water's Stillness")
-	.setCharges(15)
-	.setStagger(-2)
-	.setModifiers({ name: "Vigilance", stacks: 0 }, scalingSwiftness(2))
-	.setScalings({ critBonus: 2 });
