--- source/gear/wolfring-accurate.js
@@ -1,14 +1,0 @@
-const { GearTemplate } = require('../classes');
-
-module.exports = new GearTemplate("Accurate Wolf Ring",
-	[
-		["Passive", "Increase your Max HP by @{percentMaxHP}% and your Crit Rate by @{percentCritRate}% multiplicatively"]
-	],
-	"Trinket",
-	"Unaligned"
-).setCost(350)
-	.setSidegrades("Powerful Wolf Ring", "Swift Wolf Ring")
-	.setScalings({
-		percentMaxHP: 20,
-		percentCritRate: 10
-	});
